+++ {{cookiecutter.repo_name}}/project/mods_available/site_metas/settings.py
@@ -2,4 +2,7 @@
 TEMPLATES[0]['OPTIONS']['context_processors'] = add_to_tuple(
     TEMPLATES[0]['OPTIONS']['context_processors'],
     'project.utils.context_processors.site_metas',
+    'project.utils.context_processors.marketingtags_metas',
 )
+
+print TEMPLATES[0]['OPTIONS']['context_processors']